signUp 리덕스 State 업데이트
