+++ frontend/src/components/GuestPage/GuestPage.module.css
@@ -1,5 +1,5 @@
 .wrapper {
-  background-image: url(../assets/MeshBackground.png);
+  background-image: url(../../assets/MeshBackground.png);
   background-repeat: no-repeat;
   background-size: cover;
 }
